fix TableConfig constants
